% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffexpr.R
\name{diffexpr}
\alias{diffexpr}
\title{diffexpr function}
\usage{
diffexpr(genes1, genes2)
}
\arguments{
\item{genes1}{genes from the first list}

\item{genes2}{genes from the second list}
}
\value{
the difference value
}
\description{
Makes the difference on two lists of genes input
}
