% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneexpr.default.R
\name{geneexpr.default}
\alias{geneexpr.default}
\title{geneepxr.default function}
\usage{
\method{geneexpr}{default}(values, names)
}
\arguments{
\item{values}{a specific value}

\item{names}{a specific name}
}
\value{
a list of the results of class "geneexpr"
}
\description{
S3 default function of geneexpr
}
