% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.geneexpr.R
\name{plot.geneexpr}
\alias{plot.geneexpr}
\title{plot.geneexpr function}
\usage{
\method{plot}{geneexpr}(object)
}
\arguments{
\item{object}{geneexpr object}
}
\value{
a histogram of the results of class "geneexpr"
}
\description{
S3 plotting function of geneexpr
}
