% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.geneexpr.R
\name{print.geneexpr}
\alias{print.geneexpr}
\title{print.geneexpr function}
\usage{
\method{print}{geneexpr}(object)
}
\arguments{
\item{object}{geneexpr object}
}
\value{
prints a list of the results of class "geneexpr"
}
\description{
S3 printing function of geneexpr
}
