% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{skewness}
\alias{skewness}
\title{Skewness function}
\usage{
skewness(a, na.rm = FALSE)
}
\arguments{
\item{a}{enter values}

\item{na.rm}{If TRUE, will remove NA values}
}
\value{
the results
}
\description{
Calculate the skewness
}
