% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.geneexpr.R
\name{summary.geneexpr}
\alias{summary.geneexpr}
\title{summary.geneexpr function}
\usage{
\method{summary}{geneexpr}(object)
}
\arguments{
\item{object}{geneexpr object}
}
\value{
the summary of the results of class "geneexpr"
}
\description{
S3 summary function of geneexpr
}
